agent-ui: move constants to a dedicated folder
