+++ packages/agent-ui/src/containers/segmentPage.test.js
@@ -5,7 +5,7 @@ import chai, { expect } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 
-import * as statusTypes from '../status';
+import * as statusTypes from '../constants/status';
 
 import { SegmentPage, mapStateToProps } from './segmentPage';
