+++ packages/agent-ui/src/containers/topBar.js
@@ -5,7 +5,7 @@ import { withRouter, NavLink, Route } from 'react-router-dom';
 
 import { AppendSegmentButton, CreateMapButton } from '../components';
 
-import { openCreateMapDialog } from '../actions/createMap';
+// import { openCreateMapDialog } from '../actions/createMap';
 
 const renderTopBarLinks = path => {
   const parts = path.split('/').filter(p => p);
@@ -75,5 +75,5 @@ function mapStateToProps(state, ownProps) {
 }
 
 export default withRouter(
-  connect(mapStateToProps, { mapDialog: openCreateMapDialog })(TopBar)
+  connect(mapStateToProps, { mapDialog: () => {} })(TopBar)
 );
