+++ packages/agent-ui/src/reducers/appendSegment.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import * as actionTypes from '../actions/actionTypes';
-import * as statusTypes from '../status';
+import * as actionTypes from '../constants/actionTypes';
+import * as statusTypes from '../constants/status';
 
 const appendSegmentDialogReducer = (state = { show: false }, action) => {
   switch (action.type) {
