+++ packages/agent-ui/src/reducers/appendSegment.test.js
@@ -2,8 +2,8 @@ import { expect } from 'chai';
 
 import appendSegment from './appendSegment';
 
-import * as actionTypes from '../actions/actionTypes';
-import * as statusTypes from '../status';
+import * as actionTypes from '../constants/actionTypes';
+import * as statusTypes from '../constants/status';
 
 describe('appendSegment reducer', () => {
   const openedState = {
