+++ packages/agent-ui/src/reducers/segment.test.js
@@ -2,8 +2,8 @@ import { expect } from 'chai';
 
 import segment from './segment';
 
-import * as actionTypes from '../actions/actionTypes';
-import * as statusTypes from '../status';
+import * as actionTypes from '../constants/actionTypes';
+import * as statusTypes from '../constants/status';
 
 describe('segment reducer', () => {
   const requestSegmentAction = {
