+++ packages/agent-ui/src/reducers/segments.test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 
 import segments from './segments';
-import * as actionTypes from '../actions/actionTypes';
-import * as statusTypes from '../status';
+import * as actionTypes from '../constants/actionTypes';
+import * as statusTypes from '../constants/status';
 
 describe('segments reducer', () => {
   it('returns previous state for unknown action', () => {
